refactor: rename oxford_spires_utils to oxspires_tools
