+++ oxspires_tools/cpp/include/oxford_spires_utils/progress_bar.h
