+++ oxspires_tools/cpp/main.cpp
