+++ oxspires_tools/cpp/src/filter_pcd_with_bt.cpp
