+++ oxspires_tools/cpp/src/get_occ_free_from_bt.cpp
