+++ oxspires_tools/cpp/src/progress_bar.cpp
