+++ oxspires_tools/trajectory/README.md
