+++ oxspires_tools/trajectory/file_interfaces/base.py
