+++ oxspires_tools/trajectory/json_handler_submap_utils.py
