+++ oxspires_tools/trajectory/nerf_json_handler.py
