+++ oxspires_tools/trajectory/pose_convention.py
