+++ oxspires_tools/utils.py
@@ -5,9 +5,9 @@ import open3d as o3d
 from pypcd4 import PointCloud
 from tqdm import tqdm
 
-from oxford_spires_utils.point_cloud import convert_e57_to_pcd, transform_3d_cloud
-from oxford_spires_utils.se3 import se3_matrix_to_xyz_quat_wxyz, xyz_quat_wxyz_to_se3_matrix
-from oxford_spires_utils.trajectory.pose_convention import PoseConvention
+from oxspires_tools.point_cloud import convert_e57_to_pcd, transform_3d_cloud
+from oxspires_tools.se3 import se3_matrix_to_xyz_quat_wxyz, xyz_quat_wxyz_to_se3_matrix
+from oxspires_tools.trajectory.pose_convention import PoseConvention
 
 
 def convert_e57_folder_to_pcd_folder(e57_folder, pcd_folder):
